chore: move swap tests into dedicated directory
