+++ apps/wallet-cli/src/test/commands/swap/quote.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, beforeAll, afterAll, afterEach } from "bun:test";
-import { MockServer } from "../helpers/mock-server";
-import { runCli } from "../helpers/cli-runner";
-import { makeSessionDir } from "../helpers/session-fixture";
-import { ETH_ADDRESS, ETH_DESCRIPTOR } from "../helpers/constants";
+import { MockServer } from "../../helpers/mock-server";
+import { runCli } from "../../helpers/cli-runner";
+import { makeSessionDir } from "../../helpers/session-fixture";
+import { ETH_ADDRESS, ETH_DESCRIPTOR } from "../../helpers/constants";
 
 /** Minimal `RawQuote` row for `/quote` — enough for `normalizeQuote` + `buildQuoteDetails`. */
 const MOCK_QUOTE_ROW = {
